+++ .github/actions/ci-action/entrypoint.sh
@@ -1,2 +1,3 @@
 #!/bin/bash
+py3
 make ci