+++ scripts/assume-role-s3-copy.py
@@ -1,7 +1,7 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 import sys, os, logging, argparse, boto3, socket, textwrap
 from yaml import safe_load
-from urlparse import urlparse
+from urllib.parse import urlparse
 from socket import gethostname as ghn
 
 __version__ = '0.1'