+++ scripts/assume_role_lib/sts.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 import sys, os, boto3
 from .util import unwrap